Remove home and about pages
