+++ package.json
@@ -14,14 +14,16 @@
     "@emotion/react": "^11.13.0",
     "@emotion/styled": "^11.13.0",
     "@fontsource/roboto": "^5.1.0",
-    "@mui/icons-material": "^5.16.7",
+    "@mui/icons-material": "^5.17.1",
     "@mui/material": "^5.16.7",
     "@mui/x-charts": "^7.13.0",
     "@mui/x-data-grid": "^7.21.0",
     "@types/jsdom": "^21.1.7",
+    "axios": "^1.8.4",
     "dotenv": "^16.4.5",
     "jsdom": "^25.0.1",
     "lodash.debounce": "^4.0.8",
+    "motion": "^12.6.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router": "^7.2.0",
