+++ src/app.css
@@ -94,7 +94,7 @@ select {
   color: #18a0fb;
 }
 
-.icon_dark_mode_colors {
+/* .icon_dark_mode_colors {
   filter: invert(94%) sepia(100%) saturate(25%) hue-rotate(13deg)
     brightness(105%) contrast(108%);
 }
@@ -102,7 +102,7 @@ select {
 .icon_light_mode_colors {
   filter: invert(0%) sepia(0%) saturate(7500%) hue-rotate(327deg)
     brightness(96%) contrast(104%);
-}
+} */
 
 .containers_dark_colors {
   background-color: rgba(33, 33, 33, 0.51);
@@ -124,7 +124,7 @@ select {
 @media screen and (min-width: 1200px) {
   .app-main-container {
     flex-direction: row;
-
+    width: 100%;
     min-height: max-content;
     height: 100vh;
   }
@@ -133,7 +133,6 @@ select {
   }
   .app-section-container {
     width: 75%;
-
     overflow-y: auto;
   }
 }
