+++ src/components/Chart.css
@@ -1,5 +1,3 @@
 .line_chart_container {
   width: 100%;
-  /* height: 23.375rem;
-  padding: 2.375rem; */
 }
