+++ src/components/DashBoardTitle/DashboardTitle.css
@@ -1,8 +1,8 @@
 .dashboard_title_container {
-  width: 98%;
-  height: 1.875rem;
-  /* margin: 2.5rem auto 1.938rem auto; */
-  margin: 1rem auto;
+  width: 100%;
+  height: 4rem;
+  margin: 0 auto;
+  padding: 1rem;
   display: flex;
   align-items: center;
   justify-content: space-between;
@@ -25,7 +25,11 @@
   .dashboard_title_container {
     /* position: fixed; */
     /* margin-top: 100px; */
-    margin-left: 20px;
-    width: 98%;
+    /* margin-left: 20px; */
+
+    padding: 1rem 1rem;
+    height: 4rem;
+    /* margin: auto;
+    width: 100%; */
   }
 }
