+++ src/components/DisplayStockOverview/DisplayStockOverview.tsx
@@ -1,10 +1,11 @@
-import { useContext } from "react";
-import { stockContext } from "../../App";
 import "./DisplayStockOverview.css";
 import arrow from "../../images/stockOverview/tabler_arrow-narrow-down.svg";
+import { useStocksContext } from "../StocksContextProvider";
+import { Skeleton } from "@mui/material";
 
 export default function DisplayStockOverview() {
-  const { stockOverview, stockDetails } = useContext(stockContext);
+  const { stockOverview, stockDetails, isStocksInfoLoading } =
+    useStocksContext();
   const { symbol = "", price = "", changePercent = "" } = stockOverview || {};
   const isNegativeChange = +changePercent < 0;
 
@@ -12,33 +13,50 @@ export default function DisplayStockOverview() {
     <section className="stock_overview_container">
       <div className="logo_symbol_container">
         <span className="stock_overview_item">
-          {stockDetails ? (
-            <img className="logo" src={stockDetails.logo} alt="company logo" />
+          {!isStocksInfoLoading ? (
+            <img className="logo" src={stockDetails?.logo} alt="company logo" />
           ) : (
-            ""
+            <Skeleton variant="circular" width={57} height={57} />
           )}
         </span>
         <div className="symbol_name_container">
-          <span className="stock_overview_name ">{stockDetails?.name}</span>
-          <span className="stock_overview_symbol">{symbol}</span>
+          <span className="stock_overview_name ">
+            {!isStocksInfoLoading ? (
+              stockDetails?.name
+            ) : (
+              <Skeleton variant="rounded" width={210} height={30} />
+            )}
+          </span>
+          <span className="stock_overview_symbol">
+            {!isStocksInfoLoading ? (
+              symbol
+            ) : (
+              <Skeleton variant="rounded" width={210} height={20} />
+            )}
+          </span>
         </div>
       </div>
-
       <div className="percent_price_update_container">
-        <div className="percent_price_container">
-          <div className="change_percent_container">
-            <span className="place_holder"></span>
-            <span className="percent_text">{`${changePercent}%`}</span>
+        {!isStocksInfoLoading ? (
+          <>
+            <div className="percent_price_container">
+              <div className="change_percent_container">
+                <span className="place_holder"></span>
+                <span className="percent_text">{`${changePercent}%`}</span>
 
-            <img
-              src={arrow}
-              className={`${isNegativeChange ? "percent_arrow_down" : ""}`}
-            />
-          </div>
+                <img
+                  src={arrow}
+                  className={`${isNegativeChange ? "percent_arrow_down" : ""}`}
+                />
+              </div>
 
-          <span className="price">{`$${price}`}</span>
-        </div>
-        <div className="last_updated"></div>
+              <span className="price">{`$${price}`}</span>
+            </div>
+            <div className="last_updated"></div>{" "}
+          </>
+        ) : (
+          <Skeleton variant="rounded" width={150} height={23} />
+        )}
       </div>
     </section>
   );
