+++ src/components/NavBar/NarBar.tsx
@@ -1,12 +1,22 @@
+import RecentlySeenStock from "../RecentlySeenStock/RecentlySeenStock";
+import { useStocksContext } from "../StocksContextProvider";
 import "./NavBar.css";
-import { useContext } from "react";
-import { stockContext } from "../../App";
 
 export default function NavBar() {
-  const { primaryColors, accentColors } = useContext(stockContext);
+  const { primaryColors, accentColors, recentlySeenStocks } =
+    useStocksContext();
   return (
     <nav className={`navbar_container ${primaryColors}`}>
-      <h1 className={`navbar_logo colors ${accentColors}`}> Stock Dashboard</h1>
+      <div className="navbar_logo_container">
+        <h1 className={`navbar_logo colors ${accentColors}`}>
+          {" "}
+          Stock Dashboard
+        </h1>
+      </div>
+
+      {recentlySeenStocks.map((stock) => (
+        <RecentlySeenStock key={stock.symbol} stock={stock} />
+      ))}
     </nav>
   );
 }
