+++ src/components/RecentlySeenStock/RecentlySeenStock.tsx
@@ -1,43 +1,27 @@
 import "./RecentlySeenStock.css";
-import arrow from "../../images/stockOverview/tabler_arrow-narrow-down.svg";
-import { RecentlySeenStocks, stockContext } from "../../App";
-import { useContext } from "react";
+import { useStocksContext } from "../StocksContextProvider";
+import { RecentlySeenStocks } from "../../types";
+import DeleteForeverSharpIcon from "@mui/icons-material/DeleteForeverSharp";
 
 export default function RecentlySeenStock({
   stock,
 }: {
   stock: RecentlySeenStocks;
 }) {
-  const { symbol, price, changePercent, logo, name } = stock;
-  const { deleteToRecentlySeenStocks } = useContext(stockContext);
-  const isNegativeChange = +changePercent < 0;
+  const { price, logo, name } = stock;
+  const { deleteToRecentlySeenStocks, primaryColors, secondaryColors } =
+    useStocksContext();
+
   return (
-    <section className="recent_stock_overview_container">
-      <div className="logo_symbol_container">
-        <span className="stock_overview_item">
-          <img className="logo" src={logo} alt="company logo" />
-        </span>
-        <div className="symbol_name_container">
-          <span className="stock_overview_name ">{name}</span>
-          <span className="stock_overview_symbol">{symbol}</span>
-        </div>
-      </div>
+    <button className={`recent_stock_overview_container ${secondaryColors}`}>
+      <img className="recent_logo" src={logo} alt="company logo" />
+      <h4 className="recent_name ">{name}</h4>
+      <h4 className="recent_price">{`$${price}`}</h4>
 
-      <div className="percent_price_update_container">
-        <div className="percent_price_container">
-          <div className="change_percent_container">
-            <span className="place_holder"></span>
-            <span className="percent_text">{`${changePercent}%`}</span>
-
-            <img
-              src={arrow}
-              className={`${isNegativeChange ? "percent_arrow_down" : ""}`}
-            />
-          </div>
-          <span className="price">{`$${price}`}</span>
-          <button onClick={() => deleteToRecentlySeenStocks(stock)}>X</button>
-        </div>
-      </div>
-    </section>
+      <DeleteForeverSharpIcon
+        sx={{ color: `${primaryColors}` }}
+        onClick={() => deleteToRecentlySeenStocks(stock)}
+      />
+    </button>
   );
 }
