+++ src/components/SearchBar/SearchBar.css
@@ -10,6 +10,5 @@
   .searchbar_container {
     height: 5.5rem;
     width: 100%;
-    /* position: fixed; */
   }
 }
