+++ src/components/StockSearchInput/StockSearchInput.css
@@ -1,6 +1,5 @@
 .search_bar_container {
   width: 24.4375rem;
-  margin: 0 auto;
   padding: 0.6875rem 1.3125rem;
   margin: 1.8125rem 1.25rem;
 }
