+++ src/helpers/fetchStockDetails.ts
@@ -1,22 +1,7 @@
+import { StockDetails } from "../types";
 import fetchFinnhubStockData from "./fetchFinnhubStockData";
 const FINNHUB_API_KEY = import.meta.env.VITE_FINNHUB_API_KEY;
 const FINNHUB_BASE_URL = "https://finnhub.io/api/v1";
-
-export interface StockDetails {
-  country: string;
-  currency: string;
-  estimateCurrency: string;
-  exchange: string;
-  finnhubIndustry: string;
-  ipo: string;
-  logo: string;
-  marketCapitalization: number;
-  name: string;
-  phone: string;
-  shareOutstanding: number;
-  ticker: string;
-  weburl: string;
-}
 
 export default async function fetchStockDetails(
   stockSymbol: string
