+++ src/helpers/organizedStocksInGroups.ts
@@ -1,5 +1,5 @@
+import { StocksData } from "../types";
 import categorizeByTimePeriod from "./categorizeByTimePeriod";
-import { StocksData } from "../App";
 import { DailyStocksApiResponse } from "./fetchVantageStockData";
 
 export enum GroupType {
