+++ src/pages/NoData.css
@@ -1,0 +1,10 @@
+.noData_container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 98%;
+  height: 85vh;
+  margin: auto;
+  border-radius: 2.375rem;
+}
